refactor: move multipart DTO contracts
